docker/producer/requirements.txt
kafka-python
python-twitter
tweepy
python-decouple